feat: release in remove within controllers and seed

+++ src/controllers/products-controllers.ts
@@ -1,5 +1,6 @@
 import { NextFunction, Request, Response } from 'express'
 import { knex } from '@/database/knex'
+import { AppError } from '@/utils/AppError'
 import { z } from 'zod'
 
 class ProductController {
@@ -50,6 +51,15 @@ class ProductController {
 
       const { name, price } = bodySchema.parse(request.body)
 
+      const product = await knex<ProductRepository>('products')
+        .select()
+        .where({ id })
+        .first()
+
+      if (!product) {
+        throw new AppError('Product not found')
+      }
+
       await knex<ProductRepository>('products')
         .update({ name, price, updated_at: knex.fn.now() })
         .where({ id })
@@ -60,7 +70,7 @@ class ProductController {
     }
   }
 
-  async delete(request: Request, response: Response, next: NextFunction) {
+  async remove(request: Request, response: Response, next: NextFunction) {
     try {
       const id = z
         .string()
@@ -69,6 +79,15 @@ class ProductController {
           message: 'id must be a number',
         })
         .parse(request.params.id)
+
+      const product = await knex<ProductRepository>('products')
+        .select()
+        .where({ id })
+        .first()
+
+      if (!product) {
+        throw new AppError('Product not found')
+      }
 
       await knex<ProductRepository>('products').delete().where({ id })
